+++ src/app/(profile-related)/profile/_components/confirmation.tsx
@@ -3,9 +3,10 @@
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { useEffect, useRef, useState } from "react";
-import { deleteUserAction } from "../actions";
+import { deleteUserAction } from "../../../_actions/user";
 import { signOut } from "next-auth/react";
 import { AlertTriangle } from "lucide-react";
+import { catchError } from "@/lib/utils";
 
 interface DeleteConfirmationProps {
   setWillDelete: React.Dispatch<React.SetStateAction<boolean>>;
@@ -37,7 +38,7 @@ export default function DeleteConfirmation({
         callbackUrl: "/",
       });
     } catch (err) {
-      console.log(err);
+      catchError(err);
     } finally {
       setIsLoading(false);
     }
