+++ src/config/languages.ts
@@ -1,0 +1,13 @@
+export const snippetLanguages: { label: string; value: string }[] = [
+  { label: "C++", value: "c++" },
+  { label: "C#", value: "c#" },
+  { label: "Go", value: "go" },
+  { label: "HTML", value: "html" },
+  { label: "Java", value: "java" },
+  { label: "Javascript", value: "javascript" },
+  { label: "PHP", value: "php" },
+  { label: "Python", value: "python" },
+  { label: "Ruby", value: "ruby" },
+  { label: "Swift", value: "swift" },
+  { label: "Typescript", value: "typescript" },
+];
